requirements.txt
absl-py==0.15.0
altgraph==0.17.2
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
flatbuffers==1.12
fonttools==4.34.4
future==0.18.2
gast==0.3.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
librosa==0.6.3
llvmlite==0.31.0
lxml==4.9.1
Markdown==3.4.1
matplotlib==3.5.2
Nuitka==1.0.1
numba==0.48.0
numpy==1.19.2
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pefile==2022.5.30
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
py2exe==0.11.1.1
pyarmor==7.6.0
pyarmor-webui==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyMySQL==1.0.2
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
reportlab==3.5.67
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-addons==0.12.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tinui==3.23.0
tinyaes==1.0.3
tqdm==4.63.0
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.10
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.1
